requirements.txt
Django==5.1.1
langchain==0.3.2
langchain_community==0.3.1
langchain_core==0.3.9
langchain_google_genai==2.0.0
python-dotenv==1.0.1
